pulled display util funcs out, cleaned arg parsing
