+++ examples/franka_example.py
@@ -6,25 +6,12 @@
 
 import argparse
 
-import cv2
-import numpy as np
+from display_utils import display_rgb
 
 import habitat
 from habitat.core.logging import logger
 
 SENSOR_KEY = "rgb"
-
-
-def display_grayscale(image):
-    img_bgr = np.repeat(image, 3, 2)
-    cv2.imshow("Depth Sensor", img_bgr)
-    cv2.waitKey(0)
-
-
-def display_rgb(image):
-    img_bgr = image[..., ::-1]
-    cv2.imshow("RGB", img_bgr)
-    cv2.waitKey(0)
 
 
 def example(render):
